feat(vim) replace linter plugin
While I have sworn by Ale before now, it does not work that well with
vim in Git Bash on Windows. So far as I can tell, it does not recognize
the tsconfig.json files for a project when it is not in the root folder,
and I'm done trying to figure that.

- Replace ale with coc.vim.
- Disable supertab while I assess coc.vim.
- Remove vim-searchant, as I haven't used it in months.