nvim: migrate to Neovim LSP
This one was a while coming becasue I kept putting it of and putting it
off and putting it off...for two years. The changes in nvim 0.11 make it
far easier to setup, which became my prompt to do so.

There are many changes still to be made, but I'm satisfied with this as
a core migration.