vim: adjust nvim-cmp to favour the buffer
My work project is a ginormous monorepo and as a consequence it bogs
down everything LSP. I enjoyed the text matches provided by coc.nvim, so
I've weighted completions to favour the buffer. I *will* keep an this
with the intent of tweaking it to balance performance + results.

I value low latency when I work with vim, so I feel any lag or delay in
a keen way.